+++ lib/stealth/dispatcher.rb
@@ -36,16 +36,7 @@ module Stealth
       end
 
       bot_controller = BotController.new(service_message: service_message)
-
-      if ignore_message?(session: bot_controller.current_session)
-        Stealth::Logger.l(
-          topic: 'user',
-          message: "User #{service_message.sender_id}; ignoring message due to session (#{bot_controller.current_session})"
-        )
-        return false
-      else
-        bot_controller.route
-      end
+      bot_controller.route
     end
 
     private
@@ -73,10 +64,6 @@ module Stealth
           topic: 'user',
           message: "User #{service_message.sender_id} -> #{message}"
         )
-      end
-
-      def ignore_message?(session:)
-        %w[catch_all interrupt].include?(session.flow_string)
       end
   end
 end